[ADD] Combined util files and made util for lat_lon_arr
